Video 2

+++ src/controller/UserController.ts
@@ -1,4 +1,4 @@
-import {getRepository} from "typeorm";
+import {EntityNotFoundError, getRepository} from "typeorm";
 import {NextFunction, Request, Response} from "express";
 import {User} from "../entity/User";
 
@@ -20,6 +20,7 @@ export class UserController {
 
   async remove(request: Request, response: Response, next: NextFunction) {
     let userToRemove = await this.userRepository.findOne(request.params.id);
+    if (!userToRemove) throw new Error('User not found');
     await this.userRepository.remove(userToRemove);
   }
 }
